+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `continueWithAutosave` to forms validation
+- `generatePdfFromSubmissionData()` function to `PDF` class
 
 ## [5.0.1] - 2023-11-01